Project_10/azure-app_deployment/requirements.txt
azure-functions
numpy
scikit-learn
pandas